+++ android-systrace-plugin/src/main/java/com/androidperf/systrace/TraceFixMethodTracer.java
@@ -1,6 +1,5 @@
 package com.androidperf.systrace;
 
-import com.androidperf.systrace.method.TraceMethod;
 import com.androidperf.systrace.tools.Constants;
 
 import org.objectweb.asm.ClassVisitor;
@@ -9,144 +8,139 @@ import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.commons.AdviceAdapter;
 
-import java.util.HashMap;
-import java.util.concurrent.ConcurrentHashMap;
-
 public class TraceFixMethodTracer extends ClassVisitor {
     private final int asmApi;
-    private final HashMap<String, TraceMethod> mCollectedMethodMap;
-    private final HashMap<String, TraceMethod> mCollectedIgnoreMethodMap;
-    private final HashMap<String, TraceMethod> mCollectedBlackMethodMap;
-    private final HashMap<String, String> mCollectedClassExtendMap;
-    private final boolean isActivityOrSubClass = false;
-    private String mClassName;
-    private boolean isABSClass = false;
-    private String[] mInterfaces;
-    private boolean hasWindowFocusMethod = false;
+    private final boolean traceConstructors;
+    private final boolean traceClassInitializers;
+    private final boolean traceSyntheticMethods;
+    private final boolean traceBridgeMethods;
+    private String className;
 
     public TraceFixMethodTracer(int api, ClassVisitor cv) {
+        this(api, cv, true, true, false, false);
+    }
+
+    public TraceFixMethodTracer(
+            int api,
+            ClassVisitor cv,
+            boolean traceConstructors,
+            boolean traceClassInitializers,
+            boolean traceSyntheticMethods,
+            boolean traceBridgeMethods) {
         super(api, cv);
         this.asmApi = api;
-        this.mCollectedMethodMap = new HashMap<>();
-        this.mCollectedClassExtendMap = new HashMap<>();
-        this.mCollectedIgnoreMethodMap = new HashMap<>();
-        this.mCollectedBlackMethodMap = new HashMap<>();
+        this.traceConstructors = traceConstructors;
+        this.traceClassInitializers = traceClassInitializers;
+        this.traceSyntheticMethods = traceSyntheticMethods;
+        this.traceBridgeMethods = traceBridgeMethods;
     }
 
     @Override
     public void visit(int version, int access, String name, String signature,
                       String superName, String[] interfaces) {
-        System.out.println("TraceFixClassVisitor : visit -----> started : " + name);
         super.visit(version, access, name, signature, superName, interfaces);
-        this.mClassName = name.replace("/", ".");
-        if ((access & Opcodes.ACC_ABSTRACT) > 0 || (access & Opcodes.ACC_INTERFACE) > 0) {
-            this.isABSClass = true;
-        }
-        mCollectedClassExtendMap.put(mClassName, superName);
-        mInterfaces = interfaces;
-//        this.isActivityOrSubClass = isActivityOrSubClass(mClassName, mCollectedClassExtendMap);
+        this.className = name.replace("/", ".");
     }
 
     @Override
-    public MethodVisitor visitMethod(int access, String name, String desc, String signature, String[] exceptions) {
-        System.out.println("TraceFixClassVisitor : visitMethod -----> " + name);
-        if (isABSClass) {
-            return super.visitMethod(access, name, desc, signature, exceptions);
-        } else {
-            if (!hasWindowFocusMethod) {
-                hasWindowFocusMethod = TraceMethod.isWindowFocusChangeMethod(name, desc);
-            }
-            MethodVisitor methodVisitor = cv.visitMethod(access, name, desc, signature, exceptions);
-            methodVisitor = new AdviceAdapter(asmApi, methodVisitor, access, name, desc) {
-                private final Label traceStart = new Label();
-                private final Label traceEnd = new Label();
-                private boolean traceOpened = false;
-
-                @Override
-                protected void onMethodEnter() {
-                    String sectionName = mClassName + "." + name;
-                    System.out.println("TraceFixClassVisitor : onMethodEnter : " + sectionName);
-                    int length = sectionName.length();
-                    if (length > Constants.MAX_SECTION_NAME_LEN) {
-                        sectionName = sectionName.substring(length - Constants.MAX_SECTION_NAME_LEN);
-                    }
-                    mv.visitLdcInsn(sectionName);
-                    mv.visitMethodInsn(INVOKESTATIC, Constants.DEFAULT_TRACE_METHOD_BEAT_CLASS,
-                            Constants.DEFAULT_TRACE_METHOD_BEAT_BEGIN,
-                            "(Ljava/lang/String;)V", false);
-                    traceOpened = true;
-                    visitLabel(traceStart);
-                }
-
-                @Override
-                protected void onMethodExit(int opcode) {
-                    if (!traceOpened || opcode == ATHROW) {
-                        return;
-                    }
-                    mv.visitMethodInsn(INVOKESTATIC, Constants.DEFAULT_TRACE_METHOD_BEAT_CLASS,
-                            Constants.DEFAULT_TRACE_METHOD_BEAT_END,
-                            "()V", false);
-                }
-
-                @Override
-                public void visitMaxs(int maxStack, int maxLocals) {
-                    if (traceOpened) {
-                        Label handler = new Label();
-                        visitLabel(traceEnd);
-                        visitTryCatchBlock(traceStart, traceEnd, handler, null);
-                        visitLabel(handler);
-                        visitMethodInsn(INVOKESTATIC, Constants.DEFAULT_TRACE_METHOD_BEAT_CLASS,
-                                Constants.DEFAULT_TRACE_METHOD_BEAT_END,
-                                "()V", false);
-                        throwException();
-                    }
-                    super.visitMaxs(maxStack, maxLocals);
-                }
-            };
+    public MethodVisitor visitMethod(
+            int access,
+            String name,
+            String desc,
+            String signature,
+            String[] exceptions) {
+        MethodVisitor methodVisitor = cv.visitMethod(access, name, desc, signature, exceptions);
+        if (!shouldTraceMethod(access, name)) {
             return methodVisitor;
         }
-    }
+        return new AdviceAdapter(asmApi, methodVisitor, access, name, desc) {
+            private final Label traceStart = new Label();
+            private final Label traceEnd = new Label();
+            private boolean traceOpened = false;
 
-    @Override
-    public void visitEnd() {
-//        if (!hasWindowFocusMethod && isActivityOrSubClass && isNeedTrace) {
-//            insertWindowFocusChangeMethod(cv, className);
-//        }
-        super.visitEnd();
-    }
-
-    private void insertWindowFocusChangeMethod(ClassVisitor cv, String classname) {
-        MethodVisitor methodVisitor = cv.visitMethod(Opcodes.ACC_PUBLIC, Constants.WINDOW_FOCUS_METHOD,
-                Constants.WINDOW_FOCUS_METHOD_ARGS, null, null);
-        methodVisitor.visitCode();
-        methodVisitor.visitVarInsn(Opcodes.ALOAD, 0);
-        methodVisitor.visitVarInsn(Opcodes.ILOAD, 1);
-        methodVisitor.visitMethodInsn(Opcodes.INVOKESPECIAL,
-                Constants.ACTIVITY_CLASS,
-                Constants.WINDOW_FOCUS_METHOD,
-                Constants.WINDOW_FOCUS_METHOD_ARGS,
-                false);
-        //traceWindowFocusChangeMethod(methodVisitor, classname);
-        methodVisitor.visitInsn(Opcodes.RETURN);
-        methodVisitor.visitMaxs(2, 2);
-        methodVisitor.visitEnd();
-    }
-
-    private boolean isActivityOrSubClass(String className, ConcurrentHashMap<String, String> mCollectedClassExtendMap) {
-        className = className.replace(".", "/");
-        boolean isActivity = className.equals(Constants.ACTIVITY_CLASS)
-                || className.equals(Constants.V4_ACTIVITY_CLASS)
-                || className.equals(Constants.V7_ACTIVITY_CLASS);
-        if (isActivity) {
-            return true;
-        } else {
-            if (!mCollectedClassExtendMap.containsKey(className)) {
-                return false;
-            } else {
-                return isActivityOrSubClass(mCollectedClassExtendMap.get(className), mCollectedClassExtendMap);
+            @Override
+            protected void onMethodEnter() {
+                mv.visitLdcInsn(buildSectionName(className, name, desc));
+                mv.visitMethodInsn(INVOKESTATIC, Constants.DEFAULT_TRACE_METHOD_BEAT_CLASS,
+                        Constants.DEFAULT_TRACE_METHOD_BEAT_BEGIN,
+                        "(Ljava/lang/String;)V", false);
+                traceOpened = true;
+                visitLabel(traceStart);
             }
-        }
+
+            @Override
+            protected void onMethodExit(int opcode) {
+                if (!traceOpened || opcode == ATHROW) {
+                    return;
+                }
+                mv.visitMethodInsn(INVOKESTATIC, Constants.DEFAULT_TRACE_METHOD_BEAT_CLASS,
+                        Constants.DEFAULT_TRACE_METHOD_BEAT_END,
+                        "()V", false);
+            }
+
+            @Override
+            public void visitMaxs(int maxStack, int maxLocals) {
+                if (traceOpened) {
+                    Label handler = new Label();
+                    visitLabel(traceEnd);
+                    visitTryCatchBlock(traceStart, traceEnd, handler, null);
+                    visitLabel(handler);
+                    visitMethodInsn(INVOKESTATIC, Constants.DEFAULT_TRACE_METHOD_BEAT_CLASS,
+                            Constants.DEFAULT_TRACE_METHOD_BEAT_END,
+                            "()V", false);
+                    throwException();
+                }
+                super.visitMaxs(maxStack, maxLocals);
+            }
+        };
     }
 
+    private boolean shouldTraceMethod(int access, String name) {
+        if ((access & Opcodes.ACC_ABSTRACT) != 0 || (access & Opcodes.ACC_NATIVE) != 0) {
+            return false;
+        }
+        if (!traceSyntheticMethods && (access & Opcodes.ACC_SYNTHETIC) != 0) {
+            return false;
+        }
+        if (!traceBridgeMethods && (access & Opcodes.ACC_BRIDGE) != 0) {
+            return false;
+        }
+        if (!traceConstructors && "<init>".equals(name)) {
+            return false;
+        }
+        return traceClassInitializers || !"<clinit>".equals(name);
+    }
 
+    static String buildSectionName(String className, String methodName, String desc) {
+        String safeClassName = className == null ? "" : className;
+        String safeMethodName = methodName == null ? "" : methodName;
+        String safeDesc = desc == null ? "" : desc;
+        String readableName = safeClassName + "." + safeMethodName;
+        String identity = readableName + safeDesc;
+        String hashSuffix = "#" + Integer.toUnsignedString(identity.hashCode(), 36);
+        int maxReadableLength = Constants.MAX_SECTION_NAME_LEN - hashSuffix.length();
+        if (readableName.length() <= maxReadableLength) {
+            return readableName + hashSuffix;
+        }
+        String methodSuffix = "." + safeMethodName;
+        if (methodSuffix.length() < maxReadableLength) {
+            int classNameLength = maxReadableLength - methodSuffix.length();
+            return safePrefix(safeClassName, classNameLength) + methodSuffix + hashSuffix;
+        }
+        return safePrefix(safeMethodName, maxReadableLength) + hashSuffix;
+    }
+
+    private static String safePrefix(String value, int maxLength) {
+        if (maxLength <= 0) {
+            return "";
+        }
+        if (value.length() <= maxLength) {
+            return value;
+        }
+        int endIndex = maxLength;
+        if (endIndex > 0 && Character.isHighSurrogate(value.charAt(endIndex - 1))) {
+            endIndex--;
+        }
+        return value.substring(0, endIndex);
+    }
 }