+++ android-systrace-sample-high-local-debug/src/main/java/com/androidperformance/test/Test.java
@@ -1,14 +1,7 @@
 package com.androidperformance.test;
 
-import androidx.core.os.TraceCompat;
-
 public class Test {
-    public void test() {
-        TraceCompat.beginSection(new Throwable().getStackTrace()[0].getMethodName());
+    public int helper() {
+        return 1;
     }
-
-    public void test2() {
-        TraceCompat.endSection();
-    }
-
 }